// src/features/academic-warning/academicWarning.schema.ts
import { z } from 'zod'

export const academicWarningSchema = z.object({
  user_id: z.string(),
  caption: z.string(),
  value: z.string(),
  textColour: z.string().optional(),
  orderNo: z.number().optional(),
  warning_type: z.string().optional(),
  is_resolved: z.boolean().optional(),
  resolved_at: z.date().optional(),
})
export type AcademicWarning = z.infer<typeof academicWarningSchema>